requirements.txt
pytest
lxml
matplotlib
numpy
pyyaml
scipy
sortedcontainers
